Move the search command to the registry package.
And move it back to the top-level command.
